requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
mysqlclient
channels
channels-redis
Pillow
python-decouple
django-filter
drf-spectacular
requests
celery
redis
